Add secondary memory region per workload (#300)

+++ src/numa.hpp
@@ -41,12 +41,12 @@ bool verify_interleaved_page_placement(char* const start_addr, size_t memory_reg
                                        const NumaNodeIDs& target_nodes);
 
 // Check page locations based on given expected page locations.
-bool verify_partitioned_page_placement(char* const start_addr, size_t memory_region_size,
+bool verify_partitioned_page_placement(char* const start_addr, uint64_t memory_region_size,
                                        const PageLocations& expected_page_locations);
 
 // Check page locations. Based on the given target nodes and percentage, the expected locations are calulated first,
 // before the current page locations are compared with the expected ones.
-bool verify_partitioned_page_placement(char* const start_addr, size_t memory_region_size,
+bool verify_partitioned_page_placement(char* const start_addr, uint64_t memory_region_size,
                                        const NumaNodeIDs& target_nodes, const int32_t percentage_first_node);
 
 void log_numa_nodes(spdlog::level::level_enum log_level, const std::string& message, const NumaNodeIDs& nodes);
